+++ src/craftground/MinecraftEnv/src/main/cpp/cross_gl.h
@@ -1,0 +1,18 @@
+#pragma once
+#ifdef __APPLE__
+#define GL_SILENCE_DEPRECATION
+#include <OpenGL/OpenGL.h>
+// #include <OpenGL/gl.h>
+#include <OpenGL/gl3.h>
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtc/type_ptr.hpp>
+#else
+//    #include <GL/gl.h>
+#include <GL/glew.h>
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtc/type_ptr.hpp>
+#endif
+
+#define GL_PACK_REVERSE_ROW_ORDER_ANGLE 0x93A4 // extension
