+++ src/craftground/MinecraftEnv/src/main/cpp/cursor.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "cross_gl.h"
+void drawCursorCPU(int xPos, int yPos, int targetSizeX, int targetSizeY, int cursorWidth, int cursorHeight, GLubyte *pixels);
+void renderCursor(int mouseX, int mouseY);
+bool initCursorTexture();
