+++ src/craftground/MinecraftEnv/src/main/cpp/depth_capture.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "cross_gl.h"
+
+float* captureDepth(GLuint depthFramebufferId, int width, int height);
